refactor: extract article helpers
