+++ index.js
@@ -4,9 +4,7 @@ import StealthPlugin from 'puppeteer-extra-plugin-stealth'
 puppeteer.use(StealthPlugin())
 
 import fs from 'fs'
-import cleaner from 'clean-html'
 import Sentiment from 'sentiment'
-import { htmlToText } from 'html-to-text'
 import nlp from 'compromise'
 import absolutify from 'absolutify'
 import { JSDOM, VirtualConsole } from 'jsdom'
@@ -19,6 +17,9 @@ import keywordParser from './controllers/keywordParser.js'
 import lighthouseAnalysis from './controllers/lighthouse.js'
 import spellCheck from './controllers/spellCheck.js'
 import logger from './controllers/logger.js'
+import { autoDismissConsent } from './controllers/consent.js'
+import { buildLiveBlogSummary } from './controllers/liveBlog.js'
+import { getRawText, getFormattedText, getHtmlText, htmlCleaner } from './controllers/textProcessing.js'
 
 const require = createRequire(import.meta.url)
 
@@ -367,90 +368,6 @@ const articleParser = async function (browser, options, socket) {
   }
 
   // Attempt to auto-dismiss common consent popups/overlays across all frames
-  async function autoDismissConsent(page, consentOptions = {}) {
-    try {
-      const selectors = Array.isArray(consentOptions.selectors) ? consentOptions.selectors : []
-      const textPatterns = Array.isArray(consentOptions.textPatterns) ? consentOptions.textPatterns : []
-      const maxClicks = Number.isFinite(consentOptions.maxClicks) ? consentOptions.maxClicks : 3
-      const waitMs = Number.isFinite(consentOptions.waitAfterClickMs) ? consentOptions.waitAfterClickMs : 500
-
-      const frames = page.frames()
-      let clicks = 0
-
-      // Helper to click by selectors in a frame-like context
-      const clickSelectorsIn = async (ctx) => {
-        for (const sel of selectors) {
-          if (clicks >= maxClicks) break
-          try {
-            const el = await ctx.$(sel)
-            if (el) {
-              try { await el.evaluate(e => { try { e.scrollIntoView({ block: 'center' }) } catch {} }) } catch {}
-              await el.click({ delay: 20 })
-              clicks++
-              await page.waitForTimeout(waitMs)
-            }
-          } catch {}
-        }
-      }
-
-      // Helper to click by text patterns in a frame
-      const clickByTextIn = async (frame) => {
-        const patterns = textPatterns.map(s => String(s).toLowerCase())
-        const remaining = maxClicks - clicks
-        if (remaining <= 0 || patterns.length === 0) return
-        try {
-          const did = await frame.evaluate((patterns, remaining) => {
-            const isVisible = (el) => {
-              const rect = el.getBoundingClientRect()
-              const style = window.getComputedStyle(el)
-              return rect.width > 1 && rect.height > 1 && style.visibility !== 'hidden' && style.display !== 'none'
-            }
-            const isConsentContext = (el) => {
-              let n = el
-              const re = /(consent|cookie|privacy|gdpr)/i
-              while (n && n.nodeType === 1) {
-                const id = n.id || ''
-                const cls = (n.className && typeof n.className === 'string') ? n.className : ''
-                if (re.test(id) || re.test(cls)) return true
-                n = n.parentElement
-              }
-              return false
-            }
-            const candidates = Array.from(document.querySelectorAll('button, [role="button"], a, div[role="button"]'))
-            let count = 0
-            for (const el of candidates) {
-              if (count >= remaining) break
-              const txt = (el.innerText || el.textContent || '').trim().toLowerCase()
-              if (!txt) continue
-              if (patterns.some(p => txt.includes(p)) && isConsentContext(el)) {
-                try { if (isVisible(el)) { el.click(); count++ } } catch {}
-              }
-            }
-            return count
-          }, patterns, remaining)
-          clicks += Number(did) || 0
-          if (did) await page.waitForTimeout(waitMs)
-        } catch {}
-      }
-
-      // Pass 1: selectors in main frame, then in child frames
-      await clickSelectorsIn(page)
-      for (const f of frames) { if (clicks < maxClicks) await clickSelectorsIn(f) }
-
-      // Pass 2: text matches in frames
-      for (const f of frames) { if (clicks < maxClicks) await clickByTextIn(f) }
-
-      // If clicks may have triggered navigation, wait briefly for it to settle
-      if (clicks) {
-        try { await page.waitForNavigation({ waitUntil: 'domcontentloaded', timeout: 2000 }) } catch {}
-      }
-
-      // As a last resort, try Escape once to close modal overlays
-      try { await page.keyboard.press('Escape') } catch {}
-      if (waitMs) await page.waitForTimeout(100)
-    } catch {}
-  }
-
   if (!staticHtmlOverride && jsEnabled && options.consent && options.consent.autoDismiss) {
     await autoDismissConsent(page, options.consent)
   }
@@ -841,102 +758,6 @@ log('analyze', 'Evaluating meta tags')
   // Legacy readability prep removed; using structured/heuristic detection instead
 
   // Generic live-blog detector: build a concise summary from timestamped updates
-  function buildLiveBlogSummary(document) {
-    try {
-      const MAX_UPDATES = 40
-      const items = []
-      const seen = new Set()
-      const getAncestor = (node) => {
-        let n = node
-        let depth = 0
-        while (n && depth < 5) {
-          if (['ARTICLE','SECTION','LI','DIV'].includes(n.tagName)) return n
-          n = n.parentElement
-          depth++
-        }
-        return node && node.parentElement ? node.parentElement : null
-      }
-      const text = (el) => (el && el.textContent ? el.textContent.replace(/\s+/g,' ').trim() : '')
-      const times = Array.from(document.querySelectorAll('time, [datetime]')).slice(0, 200)
-      for (const t of times) {
-        const container = getAncestor(t)
-        if (!container) continue
-        if (seen.has(container)) continue
-        seen.add(container)
-        const ttl = container.querySelector('h1,h2,h3,h4,.headline,.title')
-        const p = container.querySelector('p')
-        const tt = text(t)
-        const hd = text(ttl)
-        const pv = text(p)
-        const score = (tt ? 1 : 0) + (hd.length > 12 ? 1 : 0) + (pv.length > 60 ? 1 : 0)
-        if (score >= 2) items.push({ time: tt, title: hd, body: pv })
-        if (items.length >= MAX_UPDATES) break
-      }
-      // Fallback: scan common live containers/posts when < 5 time-stamped entries
-      if (items.length < 5) {
-        const liveRoots = Array.from(document.querySelectorAll(
-          '.live, .live-blog, .liveblog, .timeline, .live_updates, .updates, .update, .post, [role="article"]'
-        )).slice(0, 200)
-        for (const root of liveRoots) {
-          if (!root) continue
-          if (seen.has(root)) continue
-          seen.add(root)
-          const ttl = root.querySelector('h1,h2,h3,h4,.headline,.title')
-          const p = root.querySelector('p')
-          const hd = text(ttl)
-          const pv = text(p)
-          if (pv.length > 120 || (hd.length > 15 && pv.length > 60)) {
-            items.push({ time: '', title: hd, body: pv })
-          }
-          if (items.length >= MAX_UPDATES) break
-        }
-      }
-      // AMP live-list scan
-      if (items.length < 3) {
-        try {
-          const ampLists = Array.from(document.querySelectorAll('amp-live-list')).slice(0, 5)
-          for (const lst of ampLists) {
-            const candidates = Array.from(lst.querySelectorAll('[role="article"], article, li, .update, .post')).slice(0, 50)
-            for (const c of candidates) {
-              if (seen.has(c)) continue
-              seen.add(c)
-              const ttl = c.querySelector('h1,h2,h3,h4,.headline,.title')
-              const p = c.querySelector('p')
-              const timeEl = c.querySelector('time, [datetime]')
-              const tt = text(timeEl)
-              const hd = text(ttl)
-              const pv = text(p)
-              if (pv.length > 120 || (hd.length > 15 && pv.length > 60) || (tt && pv.length > 60)) {
-                items.push({ time: tt, title: hd, body: pv })
-              }
-              if (items.length >= MAX_UPDATES) break
-            }
-            if (items.length >= MAX_UPDATES) break
-          }
-        } catch {}
-      }
-      // Accept when we have a few solid entries or two with sufficient body
-      const totalBody = items.reduce((acc, it) => acc + (it.body ? it.body.length : 0), 0)
-      const enough = (items.length >= 3) || (items.length >= 2 && totalBody >= 500)
-      if (enough) {
-        const used = items.slice(0, 5)
-        const html = ['<div class="live-summary">']
-        let chars = 0
-        for (const it of used) {
-          try { if (it.body) chars += it.body.length } catch {}
-          html.push('<div class="entry">')
-          if (it.time) html.push(`<div class="time">${it.time}</div>`)
-          if (it.title) html.push(`<div class="title">${it.title}</div>`)
-          if (it.body) html.push(`<p>${it.body}</p>`)
-          html.push('</div>')
-        }
-        html.push('</div>')
-        return { ok: true, html: html.join(''), count: used.length, chars }
-      }
-    } catch {}
-    return { ok: false }
-  }
-
   let liveOverride = null
   try {
     const live = buildLiveBlogSummary(dom.window.document)
@@ -1372,159 +1193,4 @@ log('analyze', 'Evaluating meta tags')
   }
 }
 
-/**
- * takes the article body and returns the raw text of the article
- *
- * @param {String} html - the html string to process
- *
- * @return {String} raw text of the article in lower case
- *
- */
-
-const getRawText = function (html) {
-  return new Promise(function (resolve) {
-    // Lowercase for analysis
-    const options = {
-      wordwrap: null,
-      noLinkBrackets: true,
-      ignoreHref: true,
-      ignoreImage: true,
-      tables: true,
-      uppercaseHeadings: false,
-      unorderedListItemPrefix: ''
-    }
-
-  // HTML > Text
-  let rawText = htmlToText(html, options)
-
-  // Normalise
-  rawText = nlp(rawText)
-  rawText.normalize()
-  rawText = rawText.out('text')
-
-  // Remove only square-bracketed segments that contain URL-like text
-  const containsUrlLike = (s) => {
-    if (!s) return false
-    const str = String(s)
-    if (/(?:https?:\/\/|ftp:\/\/)/i.test(str)) return true
-    if (/\bwww\.[^\s\]]+/i.test(str)) return true
-    if (/\b[\w-]+(?:\.[\w-]+)+(?:\/[\w\-._~:/?#\[\]@!$&'()*+,;=%]*)?/i.test(str)) return true
-    return false
-  }
-  rawText = rawText.replace(/\[[^\]]*\]/g, (m) => {
-    const inner = m.slice(1, -1)
-    return containsUrlLike(inner) ? ' ' : m
-  })
-  // Globally strip URLs from raw text (protocols, www, and bare domains)
-  const stripUrls = (s) => {
-    if (!s || typeof s !== 'string') return s
-    let out = s.replace(/(?:https?:\/\/|ftp:\/\/)[^\s]+/gi, ' ')
-    out = out.replace(/\bwww\.[^\s]+/gi, ' ')
-    out = out.replace(/\b(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+(?:[a-z]{2,})(?:\/[\w\-._~:/?#\[\]@!$&'()*+,;=%]*)?/gi, ' ')
-    return out
-  }
-  rawText = stripUrls(rawText)
-  rawText = rawText.replace(/\s+/g, ' ').trim()
-
-  resolve(rawText)
-  })
-}
-
-/**
- * takes the article body and the derived title and returns the formatted text of the article with links made absolute.
- *
- * @param {String} html - the body html string to process
- * @param {String} title - the title string to process
- * @param {String} baseurl - the base url of the page being scraped
- * @param {Object} options - the [htmltotext]{@link https://github.com/werk85/node-html-to-text} formatting options
- *
- * @return {String} formatted text of the article
- *
- */
-
-const getFormattedText = function (html, title, baseurl, options) {
-  return new Promise(function (resolve) {
-    if (typeof options === 'undefined') {
-      options = {
-        wordwrap: 100,
-        noLinkBrackets: true,
-        ignoreHref: true,
-        tables: true,
-        uppercaseHeadings: true,
-        linkHrefBaseUrl: baseurl
-      }
-    }
-
-    if (typeof options.linkHrefBaseUrl === 'undefined') {
-      options.linkHrefBaseUrl = baseurl
-    }
-
-    // HTML > Text
-    const text = htmlToText(html, options)
-
-    // If uppercase is set uppercase the title
-    if (options.uppercaseHeadings === true) {
-      title = title.toUpperCase()
-    }
-
-    const formattedText = title + '\n\n' + text
-
-    resolve(formattedText)
-  })
-}
-
-/**
- * takes the formatted article body text and returns the "clean" html text of the article
- *
- * @param {String} text - the formatted text string to process
- *
- * @return {String} the clean html text of the article
- *
- */
-
-const getHtmlText = function (text) {
-  return new Promise(function (resolve) {
-    // Replace windows line breaks with linux line breaks & split each line into array
-    const textArray = text.replace('\r\n', '\n').split('\n')
-    // Check length of text array (no of lines)
-    const codeLength = textArray.length
-    // Wrap each line in a span
-    textArray.forEach(function (line, index, array) {
-      if (codeLength === index) return
-      if (index === 2) line = line.trim()
-      array[index] = '<span>' + line + '</span>'
-    })
-    // Join each line back into a string
-    const htmlText = textArray.join('\n')
-
-    // return raw, formatted & html text
-    resolve(htmlText)
-  })
-}
-
-/**
- * takes a string of html and runs it through [clean-html]{@link https://github.com/dave-kennedy/clean-html}
- *
- * @param {String} html - the html to clean
- * @param {Object} options - the [clean-html options]{@link https://github.com/dave-kennedy/clean-html#options}
- *
- * @return {String} the cleaned html
- *
- */
-
-const htmlCleaner = function (html, options) {
-  return new Promise(function (resolve) {
-    if (typeof options === 'undefined') {
-      options = {
-        'add-remove-tags': ['blockquote', 'span'],
-        'remove-empty-tags': ['span'],
-        'replace-nbsp': true
-      }
-    }
-
-    cleaner.clean(html, options, function (html) {
-      resolve(html)
-    })
-  })
-}
 // Legacy title helpers removed; using controllers/titleDetector instead
